refactor: rename OAuth2 to OAuth2Provider
