+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
--   Added the OAuth2 recipe
+-   Added the OAuth2Provider recipe
 
 ## [0.12.0] - 2024-05-24
