+++ lib/ts/recipe/oauth2provider/types.ts
