+++ recipe/oauth2provider/index.d.ts
@@ -12,9 +12,6 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-"use strict";
-function __export(m) {
-    for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
-}
-exports.__esModule = true;
-__export(require("../../lib/build/recipe/oauth2"));
+export * from "../../lib/build/recipe/oauth2provider";
+import * as _default from "../../lib/build/recipe/oauth2provider";
+export default _default;
